cheri: riscv: Make entry and boot code CHERI aware
- Use CHERI register names when accessing memory.
- Save and restore CHERI registers on kernel entry.
- Find the infinite capability.
- Set bounds for various kernel variables.
- Call the CHERI initialization functions for capability relocations
  and riscv CHERI initialization.
- Access the new_vmalloc array through the GOT to get a valid
  writable capability.

+++ arch/riscv/include/asm/stacktrace.h
@@ -7,8 +7,8 @@
 #include <asm/ptrace.h>
 
 struct stackframe {
-	unsigned long fp;
-	unsigned long ra;
+	uintptr_t fp;
+	uintptr_t ra;
 };
 
 extern void notrace walk_stackframe(struct task_struct *task, struct pt_regs *regs,
